backend/pyproject.toml
[project]
name = "quiz"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.1",
    "django-filter>=25.1",
    "django-rest-framework>=0.1.0",
    "django-stubs>=5.2.0",
    "djoser>=2.3.1",
    "markdown>=3.8",
]
